+++ gamedata/shaders/r1/shared/common.h
@@ -7,19 +7,19 @@ uniform float4x4 	m_P;
 uniform float3x4	m_WV;
 uniform float4x4 	m_VP;
 uniform float4x4 	m_WVP;
-uniform half4		timers;
-uniform half4		fog_plane;
+uniform float4		timers;
+uniform float4		fog_plane;
 uniform float4		fog_params;		// x=near*(1/(far-near)), ?,?, w = -1/(far-near)
-uniform half4		fog_color;
+uniform float4		fog_color;
 uniform half3		L_sun_color;
 uniform half3		L_sun_dir_w;
 uniform half3		L_sun_dir_e;
-uniform half4		L_hemi_color;
-uniform half4		L_ambient;		// L_ambient.w = skynbox-lerp-factor
+uniform float4		L_hemi_color;
+uniform float4		L_ambient;		// L_ambient.w = skynbox-lerp-factor
 uniform float3 		eye_position;
 uniform half3		eye_direction;
 uniform half3		eye_normal;
-uniform	half4 		dt_params;
+uniform	float4 		dt_params;
 
 half3 	unpack_normal	(half3 v)	{ return 2*v-1;			}
 half3 	unpack_bx2	(half3 v)	{ return 2*v-1; 		}
